+++ src/plextui/__init__.py
@@ -1,3 +1,3 @@
 """Plex TUI prototype."""
 
-__version__ = "0.4.4"
+__version__ = "0.4.7"